Rename `Pipeline` -> `CPU` and move to new file

+++ src/detail/memory.hpp
@@ -62,7 +62,7 @@ class Memory {
         // Load a program into memory
         void load_program(const std::vector<uint32_t>& program, uint32_t start_addr = 0) {
             for (size_t i = 0; i < program.size(); ++i) {
-                this->write_word(start_addr + i * 4, program[i]);
+                this->write_word(static_cast<uint32_t>(start_addr + i * 4), program[i]);
 			}
 		}
 };
